% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{ApproxQuantile}
\alias{ApproxQuantile}
\title{Approximate Quantile Function}
\usage{
ApproxQuantile(b, x0 = 0)
}
\arguments{
\item{b}{Vector of coefficients for (x^k, x^k-1, ..., x, 1). Note a coefficient
for 1, i.e. the normalizing constant, \emph{should be} included.}

\item{x0}{Expansion point.}
}
\value{
Approximate inverse function.
}
\description{
Returns a function to evaluate the approximate inverse CDF of an exponential
family distribution of the form \eqn{\exp(\sum_{k=1}^{K}\beta_{k}x^{k})}.
}
