% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{GetDensDeriv}
\alias{GetDensDeriv}
\title{Density Derivative}
\usage{
GetDensDeriv(b, check_normalization = TRUE)
}
\arguments{
\item{b}{Vector of coefficients for (x^k, x^k-1, ..., x, 1). Note a coefficient
for 1, i.e. the normalizing constant, \emph{should be} included.}

\item{check_normalization}{If TRUE, verify that the density integrates to 1.}
}
\description{
Returns a function to evaluate the \emph{derivative of} the density of an
exponential family distribution of the form
\eqn{\exp(\sum_{k=1}^{K}\beta_{k}x^{k})}.
}
