% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizing_constant.R
\name{IsNormalized}
\alias{IsNormalized}
\title{Check Normalization}
\usage{
IsNormalized(b, tol = 1e-08)
}
\arguments{
\item{b}{Vector of coefficients for (x^k, x^k-1, ..., x, 1). Note a coefficient
for 1, i.e. the normalizing constant, \emph{should be} included.}

\item{tol}{Tolerance.}
}
\value{
Logical indicating whether the distribution integrates to 1.
}
\description{
Checks if a distribution integrates to 1.
}
