% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizing_constant.R
\name{NormalizingConstant}
\alias{NormalizingConstant}
\title{Normalizing Constant}
\usage{
NormalizingConstant(b)
}
\arguments{
\item{b}{Vector of coefficients for (x^k, x^k-1, ..., x). Note a coefficient
for 1, i.e. the normalizing constant, should \emph{not} be included.}
}
\value{
Vector augmented with the log normalizing constant in final position.
}
\description{
Determines the normalizing constant for an exponential family distribution of
the form \eqn{\exp(\sum_{k=1}^{K}\beta_{k}x^{k})}.
}
\examples{
b_in <- c(-0.5, 0)
b_out <- NormalizingConstant(b_in)
}
