% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{rExpFam}
\alias{rExpFam}
\title{Sample Exponential Family}
\usage{
rExpFam(b, n, is_normalized = TRUE, xmin = -2, xmax = 2)
}
\arguments{
\item{b}{Vector of coefficients for (x^k, x^k-1, ..., x). If normalized,
include the normalization constant as well.}

\item{n}{Number of samples.}

\item{is_normalized}{Has the normalizing constant been included in the
coefficient vector? Default: TRUE}

\item{xmin}{Lower bound on interval to search when inverting the CDF.}

\item{xmax}{Upper bound on interval to search when inverting the CDF.}
}
\value{
Numeric vector of samples from the distribution.
}
\description{
Sample Exponential Family
}
\examples{
# Draw from standard normal distribution.
x <- rExpFam(b = c(-0.5, 0.0, -0.5 * log(2 * pi)), n = 10)
}
