% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{GetCDF}
\alias{GetCDF}
\title{Cumulative Distribution Function}
\usage{
GetCDF(b, lower_tail = TRUE)
}
\arguments{
\item{b}{Vector of coefficients for (x^k, x^k-1, ..., x, 1). Note a coefficient
for 1, i.e. the normalizing constant, \emph{should be} included.}

\item{lower_tail}{Logical. If TRUE, integrates (-Inf, x). If FALSE,
integrates (x, Inf).}
}
\value{
Function that evalutes the distribution.
}
\description{
Returns a function to evaluate the cumulative distribution function of an
exponential family distribution of the form
\eqn{\exp(\sum_{k=1}^{K}\beta_{k}x^{k})}.
}
