% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{GetDensDeriv}
\alias{GetDensDeriv}
\title{Density Derivative}
\usage{
GetDensDeriv(b)
}
\arguments{
\item{b}{Vector of coefficients for (x^k, x^k-1, ..., x, 1). Note a coefficient
for 1, i.e. the normalizing constant, \emph{should be} included.}
}
\value{
Function that evalutes the distribution.
}
\description{
Returns a function to evaluate the \emph{derivative of} the density of an
exponential family distribution of the form
\eqn{\exp(\sum_{k=1}^{K}\beta_{k}x^{k})}.
}
